% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaddisonProjectData.R
\docType{data}
\name{MaddisonCountries}
\alias{MaddisonCountries}
\title{Maddison Project data}
\format{
\subsection{\code{MaddisonData}}{

A data frame with 4 columns:
\describe{
\item{countrycode}{3-letter ISO country code}
\item{year}{Year}
\item{gdppc}{Real Gross Domestic Product (GDP) per capita in 2011$ at
\href{https://en.wikipedia.org/wiki/Purchasing_power_parity}{purchasing power parity (PPP)}
}
\item{pop}{population}
}
}

\subsection{\code{MaddisonCountries}}{

A data frame with 3 columns:
\describe{
\item{countrycode}{3-letter ISO country code}
\item{country}{Country name used by the Maddison project}
\item{region}{Geographic region including \code{country}}
}
}
}
\usage{
MaddisonCountries
}
\description{
The
\href{https://en.wikipedia.org/wiki/Maddison_Project}{Maddison project}
collates historical economic statistics from many sources.
}
\keyword{datasets}
