% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaddisonData.R
\docType{data}
\name{MaddisonData}
\alias{MaddisonData}
\title{Maddison Project data}
\format{
\subsection{\code{MaddisonData}}{

A data frame with 4 columns:
\describe{
\item{countrycode}{3-letter ISO country code}
\item{year}{numeric year starting with year 1 CE}
\item{gdppc}{Gross domestic product (GDP) per capita in 2011 dollars at
puchasing power parity (PPP)}
\item{pop}{population}
}
}
}
\source{
\url{https://www.rug.nl/ggdc/historicaldevelopment/maddison/releases/maddison-project-database-2020?lang=en}"Groningen Growth and Development Centre"
}
\usage{
MaddisonData
}
\description{
The
\href{https://en.wikipedia.org/wiki/Maddison_Project}{Maddison project}
collates historical economic statistics from many sources.
\code{MaddisonCountries} is a \code{\link{data.frame}} of all (\code{countrycode}, \code{country},
\code{region}) combinations in those data.
}
\examples{
# Get the countrycode for a country
subset(MaddisonCountries, country=='United Kingdom', countrycode)
# Select  
str(GBR <- MaddisonData[MaddisonData$countrycode=='GBR', ])
}
\keyword{datasets}
