% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS4Maddison-package.R
\docType{package}
\name{SS4Maddison-package}
\alias{SS4Maddison}
\alias{SS4Maddison-package}
\title{SS4Maddison: State Space modeling using Maddison project data}
\description{
Maddison project data in *.rda format with functions and vignettes to illustrate how to use KFAS and other tools to model Maddison data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.r-project.org}
  \item \url{https://github.com/sbgraves237/SS4Maddison}
  \item Report bugs at \url{https://github.com/sbgraves237/SS4Maddison/issues}
}

}
\author{
\strong{Maintainer}: Spencer Graves \email{spencer.graves@effectivedefense.org} (\href{https://orcid.org/0009-0005-5387-729X}{ORCID})

}
\keyword{internal}
